pyproject.toml
[tool.poetry]
name = "nsa"
version = "0.1.0"
description = "Space Agency site"
authors = ["Vladimir Kravtsov <kvitee41@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.9.19"
Django = "4.2"
django-filer = "^3.2.3"
django-admin-sortable2 = "^2.2.4"

[tool.poetry.group.dev.dependencies]
pyright = "^1.1.389"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
